fix these also

+++ GxGDEW027C44/GxGDEW027C44.cpp
@@ -209,7 +209,7 @@ void  GxGDEW027C44::drawBitmap(int16_t x, int16_t y, const uint8_t *bitmap, int1
   {
     for (uint16_t y1 = y; y1 < y + h; y1++)
     {
-      uint16_t i = x1 / 8 + y1 * w / 8;
+      uint32_t i = x1 / 8 + uint32_t(y1) * uint32_t(w) / 8;
       uint16_t pixelcolor = (bitmap[i] & (0x80 >> x1 % 8)) ? GxEPD_WHITE  : color;
       drawPixel(x1, y1, pixelcolor);
     }
